requirements.txt
kivy
kivymd
python-bidi
arabic_reshaper